created abstract PacketEvent to reduced code duplication in PacketReceiveEvent and PacketSendEvent
