+++ src/main/java/me/stormcph/lumina/event/impl/PacketSendEvent.java
@@ -1,21 +1,9 @@
 package me.stormcph.lumina.event.impl;
 
-import me.stormcph.lumina.event.Event;
 import net.minecraft.network.packet.Packet;
 
-public class PacketSendEvent extends Event {
-
-    private Packet<?> packet;
-
+public class PacketSendEvent extends PacketEvent {
     public PacketSendEvent(Packet<?> packet) {
-        this.packet = packet;
-    }
-
-    public Packet<?> getPacket() {
-        return packet;
-    }
-
-    public void setPacket(Packet<?> packet) {
-        this.packet = packet;
+        super(packet);
     }
 }
